App => TraceAndSidebar
app will take input and switch to trace view
