+++ package.json
@@ -7,7 +7,8 @@
     "lodash": "^4.17.4",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
-    "react-scripts": "1.0.17"
+    "react-scripts": "1.0.17",
+    "reselect": "^4.0.0"
   },
   "devDependencies": {
     "gh-pages": "^1.1.0"
